Add sentry alerts for critical points of failures
