+++ apps/studio/components/interfaces/ProjectCreation/ProjectCreation.constants.ts
@@ -1,3 +1,7 @@
+import { components } from 'api-types'
+
+type DesiredInstanceSize = components['schemas']['DesiredInstanceSize']
+
 // [Joshen] Obtained from https://gist.github.com/tadast/8827699
 export const COUNTRY_LAT_LON = {
   AF: { lat: 33, lon: 65 },
@@ -252,3 +256,82 @@ export const COUNTRY_LAT_LON = {
   ZM: { lat: -15, lon: 30 },
   ZW: { lat: -20, lon: 30 },
 }
+
+export const INSTANCE_SIZE_SPECS: Record<
+  DesiredInstanceSize,
+  { label: string; ram: string; cpu: string; price: string }
+> = {
+  micro: {
+    label: 'Micro',
+    ram: '1 GB',
+    cpu: '2-core ARM',
+    price: '$0.01344/hour (~$10/month)',
+  },
+  small: {
+    label: 'Small',
+    ram: '2 GB',
+    cpu: '2-core ARM',
+    price: '$0.0206/hour (~$15/month)',
+  },
+  medium: {
+    label: 'Medium',
+    ram: '4 GB',
+    cpu: '2-core ARM',
+    price: '$0.0822/hour (~$60/month)',
+  },
+  large: {
+    label: 'Large',
+    ram: '8 GB',
+    cpu: '2-core ARM',
+    price: '$0.1517/hour (~$110/month)',
+  },
+  xlarge: {
+    label: 'XL',
+    ram: '16 GB',
+    cpu: '4-core ARM',
+    price: '$0.2877/hour (~$210/month)',
+  },
+  '2xlarge': {
+    label: '2XL',
+    ram: '32 GB',
+    cpu: '8-core ARM',
+    price: '$0.562/hour (~$410/month)',
+  },
+  '4xlarge': {
+    label: '4XL',
+    ram: '64 GB',
+    cpu: '16-core ARM',
+    price: '$1.32/hour (~$960/month)',
+  },
+  '8xlarge': {
+    label: '8XL',
+    ram: '128 GB',
+    cpu: '32-core ARM',
+    price: '$2.562/hour (~$1,870/month)',
+  },
+  '12xlarge': {
+    label: '12XL',
+    ram: '192 GB',
+    cpu: '48-core ARM',
+    price: '$3.836/hour (~$2,800/month)',
+  },
+  '16xlarge': {
+    label: '16XL',
+    ram: '256 GB',
+    cpu: '64-core ARM',
+    price: '$5.12/hour (~$3,730/month)',
+  },
+}
+
+export const AVAILABLE_SIZES: DesiredInstanceSize[] = [
+  'micro',
+  'small',
+  'medium',
+  'large',
+  'xlarge',
+  '2xlarge',
+  '4xlarge',
+  '8xlarge',
+  '12xlarge',
+  '16xlarge',
+]
