+++ apps/studio/components/interfaces/SignIn/SignInSSOForm.tsx
@@ -1,4 +1,5 @@
 import HCaptcha from '@hcaptcha/react-hcaptcha'
+import * as Sentry from '@sentry/nextjs'
 import { useQueryClient } from '@tanstack/react-query'
 import { useRef, useState } from 'react'
 import toast from 'react-hot-toast'
@@ -8,16 +9,14 @@ import { BASE_PATH } from 'lib/constants'
 import { auth, buildPathWithParams } from 'lib/gotrue'
 import { Button, Form, Input } from 'ui'
 
-const signInSchema = object({
-  email: string().email('Must be a valid email').required('Email is required'),
-})
-
 const SignInSSOForm = () => {
   const queryClient = useQueryClient()
-
-  const [captchaToken, setCaptchaToken] = useState<string | null>(null)
   const captchaRef = useRef<HCaptcha>(null)
+  const [captchaToken, setCaptchaToken] = useState<string | null>(null)
 
+  const signInSchema = object({
+    email: string().email('Must be a valid email').required('Email is required'),
+  })
   const onSignIn = async ({ email }: { email: string }) => {
     const toastId = toast.loading('Signing in...')
 
@@ -54,7 +53,8 @@ const SignInSSOForm = () => {
     } else {
       setCaptchaToken(null)
       captchaRef.current?.resetCaptcha()
-      toast.error(error.message, { id: toastId })
+      toast.error(`Failed to sign in: ${error.message}`, { id: toastId })
+      Sentry.captureMessage('[CRITICAL] Failed to sign in via SSO: ' + error.message)
     }
   }
