+++ apps/studio/components/interfaces/SignIn/SignInWithSSO.tsx
@@ -1,9 +1,10 @@
+import { Lock } from 'lucide-react'
 import Link from 'next/link'
-import { Button, IconLock } from 'ui'
+import { Button } from 'ui'
 
 const SignInWithSSO = ({ searchParams }: { searchParams?: string }) => {
   return (
-    <Button asChild block size="large" type="outline" icon={<IconLock width={18} height={18} />}>
+    <Button asChild block size="large" type="outline" icon={<Lock width={18} height={18} />}>
       <Link
         href={{
           pathname: '/sign-in-sso',
