+++ apps/studio/components/interfaces/SignIn/SignUpForm.tsx
@@ -1,12 +1,23 @@
 import HCaptcha from '@hcaptcha/react-hcaptcha'
+import { CheckCircle } from 'lucide-react'
 import { useRef, useState } from 'react'
 import toast from 'react-hot-toast'
 import * as yup from 'yup'
+import * as Sentry from '@sentry/nextjs'
 
 import { useSignUpMutation } from 'data/misc/signup-mutation'
 import { BASE_PATH } from 'lib/constants'
 import { passwordSchema } from 'lib/schemas'
-import { Alert, Button, Form, IconEye, IconEyeOff, Input } from 'ui'
+import {
+  AlertDescription_Shadcn_,
+  AlertTitle_Shadcn_,
+  Alert_Shadcn_,
+  Button,
+  Form,
+  IconEye,
+  IconEyeOff,
+  Input,
+} from 'ui'
 import PasswordConditionsHelper from './PasswordConditionsHelper'
 
 const signUpSchema = passwordSchema.shape({
@@ -29,6 +40,7 @@ const SignUpForm = () => {
       setCaptchaToken(null)
       captchaRef.current?.resetCaptcha()
       toast.error(`Failed to sign up: ${error.message}`)
+      Sentry.captureMessage('[CRITICAL] Failed to sign up: ' + error.message)
     },
   })
 
@@ -58,10 +70,14 @@ const SignUpForm = () => {
           isSubmitted ? 'opacity-100' : 'opacity-0'
         }`}
       >
-        <Alert className="w-full" withIcon variant="success" title="Check your email to confirm">
-          You've successfully signed up. Please check your email to confirm your account before
-          signing in to the Supabase dashboard. The confirmation link expires in 10 minutes.
-        </Alert>
+        <Alert_Shadcn_ variant="default">
+          <CheckCircle />
+          <AlertTitle_Shadcn_>Check your email to confirm</AlertTitle_Shadcn_>
+          <AlertDescription_Shadcn_ className="text-xs">
+            You've successfully signed up. Please check your email to confirm your account before
+            signing in to the Supabase dashboard. The confirmation link expires in 10 minutes.
+          </AlertDescription_Shadcn_>
+        </Alert_Shadcn_>
       </div>
       <Form
         validateOnBlur
