+++ apps/studio/pages/reset-password.tsx
@@ -1,7 +1,8 @@
+import Link from 'next/link'
+
 import ResetPasswordForm from 'components/interfaces/SignIn/ResetPasswordForm'
 import { ForgotPasswordLayout } from 'components/layouts'
 import { withAuth } from 'hooks'
-import Link from 'next/link'
 import type { NextPageWithLayout } from 'types'
 
 const ResetPasswordPage: NextPageWithLayout = () => {
